+++ pubspec.yaml
@@ -37,7 +37,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   image_filters:
-    path: ../
+    git: https://github.com/nikolaydymura/image_filters
   path_provider: ^2.0.11
   image: ^3.2.2
